1, Rename dependence directory "external" to "3rd_party"     for the reason that it may cause conflict with inner     bazel path; 2, Update some packages, such as rules_boost, prometheus     glog, async_grpc to a much more latest version; 3, Fix slam dependent packages to the correct status and     fix the BUILD files to make it compilable, TBD. 4, Fix bug in build script, which will cause miss module     in BUILD file.

+++ cyber/logger/async_logger.cc
@@ -58,7 +58,7 @@ void AsyncLogger::Stop() {
 }
 
 void AsyncLogger::Write(bool force_flush, time_t timestamp, const char* message,
-                        int message_len) {
+                        size_t message_len) {
   if (cyber_unlikely(state_.load(std::memory_order_acquire) != RUNNING)) {
     // std::cout << "Async Logger not running!" << std::endl;
     return;
@@ -118,7 +118,8 @@ void AsyncLogger::FlushBuffer(const std::unique_ptr<std::deque<Msg>>& buffer) {
     const bool force_flush = msg.level > 0;
     module_logger_map_.find(module_name)
         ->second->Write(force_flush, msg.ts, msg.message.data(),
-                        static_cast<int>(msg.message.size()));
+                        //static_cast<int>(msg.message.size()));
+                        msg.message.size());
     buffer->pop_front();
   }
   Flush();
